+++ semcore/tooltip/src/Tooltip.jsx
@@ -202,10 +202,10 @@ class DescriptionTooltipRoot extends TooltipRoot {
   subcomponents = [DescriptionTooltip.Trigger.displayName, DescriptionTooltip.Popper.displayName];
   defaultChildren = (title, Children, props) => (
     <>
-      <Hint.Trigger {...props} aria-haspopup='dialog'>
+      <DescriptionTooltip.Trigger {...props} aria-haspopup='dialog'>
         <Children />
-      </Hint.Trigger>
-      <Hint.Popper>{title}</Hint.Popper>
+      </DescriptionTooltip.Trigger>
+      <DescriptionTooltip.Popper>{title}</DescriptionTooltip.Popper>
     </>
   );
   handlePopperVisibleChange = (visible) => {