+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.46.0] - 2024-05-29
+
+### Added
+
+- `ReferenceStripes` component.
+- `ReferenceBackground` component (previously `ReferenceLine.Background`).
+
+### Changed
+
+- `ReferenceLine.Background` deprecated in favor of `ReferenceBackground`.
+
 ## [3.45.2] - 2024-05-31
 
 ### Changed